requirements.txt
streamlit~=1.48.1
yfinance~=0.2.65
numpy~=2.3.1
matplotlib~=3.10.3
pandas~=2.3.1
seaborn~=0.13.2
graphviz~=0.21
scipy~=1.16.0
scikit-learn~=1.7.1
pillow~=11.3.0
openai~=1.101.0
PyPDF2~=3.0.1
plotly~=6.3.0
streamlit_plotly_events~=0.0.6
gspread~=6.2.1
oauth2client~=4.1.3
stripe~=12.4.0
pygments~=2.19.2
markdown-it-py~=4.0.0